requirements.txt
python-dotenv==1.0.1
GitPython==3.1.43
python-gitlab>=4.5.0

